requirements.txt
fastapi==0.135.1
httpx==0.28.1
pydantic==2.12.5
python-dotenv==1.2.2
xhtml2pdf==0.2.17
email-validator==2.2.0
uvicorn